requirements.txt
discord.py==2.3.2
openai==0.28.0
aiohttp==3.8.5